Updated logic for deductibles
