+++ Projects/MoveFlow/Sources/Model/MovingFlowModel.swift
@@ -16,63 +16,12 @@ public struct MovingFlowModel: Codable, Equatable, Hashable {
     let maxMovingDate: String
     let suggestedNumberCoInsured: Int
     let currentHomeAddresses: [MoveAddress]
-    var quotes: [Quote]
+    let potentialHomeQuotes: [Quote]
+    var homeQuote: Quote?
+    let quotes: [Quote]
     let faqs: [FAQ]
     let extraBuildingTypes: [ExtraBuildingType]
     let changeTier: ChangeTierIntentModel?
-    init(from data: OctopusGraphQL.MoveIntentFragment) {
-        id = data.id
-        let minMovingDate = data.minMovingDate
-        let maxMovingDate = data.maxMovingDate
-        let minMovingDateDate = minMovingDate.localDateToDate
-        let maxMovingDateDate = maxMovingDate.localDateToDate
-        if let minMovingDateDate, let maxMovingDateDate {
-            if minMovingDateDate < maxMovingDateDate {
-                self.minMovingDate = minMovingDate
-                self.maxMovingDate = maxMovingDate
-            } else {
-                self.maxMovingDate = minMovingDate
-                self.minMovingDate = maxMovingDate
-            }
-        } else {
-            self.minMovingDate = data.minMovingDate
-            self.maxMovingDate = data.maxMovingDate
-        }
-        isApartmentAvailableforStudent = data.isApartmentAvailableforStudent ?? false
-        maxApartmentNumberCoInsured = data.maxApartmentNumberCoInsured
-        maxApartmentSquareMeters = data.maxApartmentSquareMeters
-        maxHouseNumberCoInsured = data.maxHouseNumberCoInsured
-        maxHouseSquareMeters = data.maxHouseSquareMeters
-
-        suggestedNumberCoInsured = data.suggestedNumberCoInsured
-        currentHomeAddresses = data.currentHomeAddresses.compactMap({
-            MoveAddress(from: $0.fragments.moveAddressFragment)
-        })
-        let quotes = data.fragments.quoteFragment.quotes
-        if !quotes.isEmpty {
-            self.quotes = data.fragments.quoteFragment.quotes.compactMap({ Quote(from: $0) })
-        } else if let homeQuotes = data.fragments.quoteFragment.mtaQuotes {
-            self.quotes = data.fragments.quoteFragment.mtaQuotes?.compactMap({ Quote(from: $0) }) ?? []
-        } else {
-            self.quotes = []
-        }
-        changeTier = {
-            if let data = data.fragments.quoteFragment.homeQuotes, !data.isEmpty {
-                return ChangeTierIntentModel.initWith(data: data)
-            }
-            return nil
-        }()
-
-        self.extraBuildingTypes = data.extraBuildingTypes.compactMap({ $0.rawValue })
-
-        var faqs = [FAQ]()
-        faqs.append(.init(title: L10n.changeAddressFaqDateTitle, description: L10n.changeAddressFaqDateLabel))
-        faqs.append(.init(title: L10n.changeAddressFaqPriceTitle, description: L10n.changeAddressFaqPriceLabel))
-        faqs.append(.init(title: L10n.changeAddressFaqRentbrfTitle, description: L10n.changeAddressFaqRentbrfLabel))
-        faqs.append(.init(title: L10n.changeAddressFaqStorageTitle, description: L10n.changeAddressFaqStorageLabel))
-        faqs.append(.init(title: L10n.changeAddressFaqStudentTitle, description: L10n.changeAddressFaqStudentLabel))
-        self.faqs = faqs
-    }
 
     init(
         id: String,
@@ -85,6 +34,7 @@ public struct MovingFlowModel: Codable, Equatable, Hashable {
         maxMovingDate: String,
         suggestedNumberCoInsured: Int,
         currentHomeAddresses: [MoveAddress],
+        potentialHomeQuotes: [Quote],
         quotes: [Quote],
         faqs: [FAQ],
         extraBuildingTypes: [ExtraBuildingType]
@@ -99,6 +49,7 @@ public struct MovingFlowModel: Codable, Equatable, Hashable {
         self.maxMovingDate = maxMovingDate
         self.suggestedNumberCoInsured = suggestedNumberCoInsured
         self.currentHomeAddresses = currentHomeAddresses
+        self.potentialHomeQuotes = potentialHomeQuotes
         self.quotes = quotes
         self.faqs = faqs
         self.extraBuildingTypes = extraBuildingTypes
@@ -146,13 +97,6 @@ struct MoveAddress: Codable, Equatable, Hashable {
     let street: String
     let postalCode: String
     let city: String?
-
-    init(from data: OctopusGraphQL.MoveAddressFragment) {
-        id = data.id
-        street = data.street
-        postalCode = data.postalCode
-        city = data.city
-    }
 }
 
 struct Quote: Codable, Equatable, Hashable {
@@ -167,164 +111,15 @@ struct Quote: Codable, Equatable, Hashable {
     let id: String
     let displayItems: [DisplayItem]
     let exposureName: String?
-
-    init(from data: OctopusGraphQL.QuoteFragment.Quote) {
-        id = UUID().uuidString
-        premium = .init(fragment: data.premium.fragments.moneyFragment)
-        startDate = data.startDate.localDateToDate?.displayDateDDMMMYYYYFormat ?? data.startDate
-        let productVariantFragment = data.productVariant.fragments.productVariantFragment
-        displayName = productVariantFragment.displayName
-        exposureName = data.exposureName
-        insurableLimits = productVariantFragment.insurableLimits.compactMap({
-            .init(label: $0.label, limit: $0.limit, description: $0.description)
-        })
-        perils = productVariantFragment.perils.compactMap({ .init(fragment: $0) })
-        documents = productVariantFragment.documents.compactMap({ .init($0) })
-        contractType = Contract.TypeOfContract(rawValue: data.productVariant.typeOfContract)
-        displayItems = data.displayItems.map({ .init($0) })
-    }
-
-    init(from data: OctopusGraphQL.QuoteFragment.MtaQuote) {
-        id = UUID().uuidString
-        premium = .init(fragment: data.premium.fragments.moneyFragment)
-        startDate = data.startDate.localDateToDate?.displayDateDDMMMYYYYFormat ?? data.startDate
-        let productVariantFragment = data.productVariant.fragments.productVariantFragment
-        displayName = productVariantFragment.displayName
-        exposureName = data.exposureName
-        insurableLimits = productVariantFragment.insurableLimits.compactMap({
-            .init(label: $0.label, limit: $0.limit, description: $0.description)
-        })
-        perils = productVariantFragment.perils.compactMap({ .init(fragment: $0) })
-        documents = productVariantFragment.documents.compactMap({ .init($0) })
-        contractType = Contract.TypeOfContract(rawValue: data.productVariant.typeOfContract)
-        displayItems = data.displayItems.map({ .init($0) })
-    }
 }
 
 struct InsuranceDocument: Codable, Equatable, Hashable {
     let displayName: String
     let url: String
-
-    init(_ data: OctopusGraphQL.ProductVariantFragment.Document) {
-        self.displayName = data.displayName
-        self.url = data.url
-    }
 }
 
 struct DisplayItem: Codable, Equatable, Hashable {
     let displaySubtitle: String?
     let displayTitle: String
     let displayValue: String
-
-    init(_ data: OctopusGraphQL.QuoteFragment.Quote.DisplayItem) {
-        displaySubtitle = data.displaySubtitle
-        displayTitle = data.displayTitle
-        displayValue = data.displayValue
-    }
-
-    init(_ data: OctopusGraphQL.QuoteFragment.MtaQuote.DisplayItem) {
-        displaySubtitle = data.displaySubtitle
-        displayTitle = data.displayTitle
-        displayValue = data.displayValue
-    }
-}
-
-extension ChangeTierIntentModel {
-    static func initWith(data: [OctopusGraphQL.QuoteFragment.HomeQuote]) -> ChangeTierIntentModel {
-
-        var currentTier: Tier?
-        var currentDeductible: Deductible?
-
-        func getFilteredTiers(
-            quotes: [OctopusGraphQL.QuoteFragment.HomeQuote]
-        ) -> [Tier] {
-            // list of all unique tierNames
-            var allTiers: [Tier] = []
-
-            var uniqueTierNames: [String] = []
-            quotes
-                .forEach({ tier in
-                    let tierNameIsNotInList = uniqueTierNames.first(where: { $0 == (tier.tierName) })?.isEmpty
-                    if tierNameIsNotInList ?? true {
-                        uniqueTierNames.append(tier.tierName)
-                    }
-                })
-
-            /* filter tiers and deductibles*/
-            uniqueTierNames.forEach({ tierName in
-                let allQuotesWithNameX = quotes.filter({ $0.tierName == tierName })
-                var allDeductiblesForX: [Deductible] = []
-
-                allQuotesWithNameX
-                    .forEach({ quote in
-                        if let deductableAmount = quote.deductible?.amount {
-                            let deductible = Deductible(
-                                deductibleAmount: .init(fragment: deductableAmount.fragments.moneyFragment),
-                                deductiblePercentage: (quote.deductible?.percentage == 0)
-                                    ? nil : quote.deductible?.percentage,
-                                subTitle: quote.deductible?.displayText,
-                                premium: .init(
-                                    optionalFragment: allQuotesWithNameX.first?.premium.fragments.moneyFragment
-                                )
-                            )
-                            allDeductiblesForX.append(deductible)
-                        }
-                    })
-
-                var displayItems: [Tier.TierDisplayItem] = []
-                allQuotesWithNameX
-                    .forEach({
-                        displayItems.append(
-                            contentsOf: $0.displayItems.map({
-                                Tier.TierDisplayItem(
-                                    title: $0.displayTitle,
-                                    subTitle: $0.displaySubtitle,
-                                    value: $0.displayValue
-                                )
-                            })
-                        )
-                    })
-
-                let FAQs: [FAQ] = [
-                    .init(title: "question 1", description: "..."),
-                    .init(title: "question 2", description: "..."),
-                    .init(title: "question 3", description: "..."),
-                ]
-
-                allTiers.append(
-                    .init(
-                        id: allQuotesWithNameX.first?.id ?? "",
-                        name: allQuotesWithNameX.first?.tierName ?? "",
-                        level: allQuotesWithNameX.first?.tierLevel ?? 0,
-                        deductibles: allDeductiblesForX,
-                        premium: .init(optionalFragment: allQuotesWithNameX.first?.premium.fragments.moneyFragment)
-                            ?? .init(amount: "0", currency: "SEK"),
-                        displayItems: displayItems,
-                        exposureName: "TODO",
-                        productVariant: .init(
-                            data: allQuotesWithNameX.first?.productVariant.fragments.productVariantFragment
-                        ),
-                        FAQs: FAQs
-                    )
-                )
-            })
-            return allTiers
-        }
-
-        let tiers = getFilteredTiers(quotes: data)
-        let first = data.first
-        let intentModel: ChangeTierIntentModel = .init(
-            activationDate: first?.startDate.localDateToDate ?? Date(),
-            tiers: tiers,
-            currentPremium: .init(
-                amount: "0",  //String(currentContract.currentAgreement.premium.amount),
-                currency: "SEK$$"  //currentContract.currentAgreement.premium.currencyCode.rawValue
-            ),
-            currentTier: currentTier,
-            currentDeductible: currentDeductible,
-            canEditTier: true
-        )
-        return intentModel
-    }
-
 }
