+++ Projects/MoveFlow/Sources/Service/DemoImplementation/MoveFlowClientDemo.swift
@@ -13,6 +13,7 @@ public class MoveFlowClientDemo: MoveFlowClient {
             maxMovingDate: "2025-06-01",
             suggestedNumberCoInsured: 2,
             currentHomeAddresses: [],
+            potentialHomeQuotes: [],
             quotes: [],
             faqs: [],
             extraBuildingTypes: []
@@ -35,6 +36,7 @@ public class MoveFlowClientDemo: MoveFlowClient {
             maxMovingDate: "2025-06-01",
             suggestedNumberCoInsured: 2,
             currentHomeAddresses: [],
+            potentialHomeQuotes: [],
             quotes: [],
             faqs: [],
             extraBuildingTypes: []
